services/analytics/requirements.txt
fastapi
uvicorn[standard]
websockets
pytest
pytest-asyncio
httpx
protobuf
redis
fakeredis
apscheduler
feedparser
vaderSentiment
rq
textblob
snscrape
pyjwt
alembic
openai>=0.27,<1
